requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-jalali-date==1.1.0
django-render-partial==0.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.3.0
sorl-thumbnail==12.10.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
